Add testing framework.
Using Bandit as a unit test framework library.
Because of this I refactored most of the collision module, there are still issues.
Add unit tests for SpatialHash and OctreeCell. These are to show how
the testing framework is used and they can be improved.

issue #3

Add test spec for OctreeCell