Improve buildbot dashboard submission.
This adds a new CMake variable (not exposed via any
options — it should be set in the initial `CMakeCache.txt`)
named `SimMedTK_SUPERBUILD_SUBPROJECT_DASHBOARD`. When set,
the superbuild will not build the "ALL" target of the SimMedTK
subproject but insteadl build the "Experimental" target, which
runs tests and submits results to CDash.

This way, buildbot superbuilds can also be configured to run
SimMedTK tests (*plus* any superbuild tests) and the results
will be submitted as two separate CDash rows.

This commit also enables testing on the superbuild and adds
a single dummy test.