Get tests running on Mac OS X.
On Mac OS X, running the tests in CTest fails currently because
SFML links to some binary-blob frameworks it includes with its
source (shudder). The path to these frameworks is specified using
`@rpath`, which – since they are not installed in an rpath – means
they are not found by default.

Setting the `DYLD_FRAMEWORK_PATH` environment variable to include
the path to the frameworks causes the tests to pass. There is
almost certainly a better way to do this.