BUG: Refactor AVMNidus example with Shaders
AVMNidus was used to test the shaders, but
was not showcasing its main usage anymore.

This commit divides the example by migrating
the Shaders test into another example (Shaders)

Also incorporated the data for this new example
in MIDAS to download it during the build.

Also fixed some shaders errors.