ENH: Rename the nonlinear solver
to just NewtonMethod. Update the solve routine
so it always goes into the armijo loop, subsequent
function norms are computed there. If the user wants
to disable the armijo updated then it should set the
armijoMax in the base class. The armijo function now
takes an extra parameter, the old function evaluation
norm.