ENH: Clean and refactor the penalty method.
Add an alias for a matrix map.
Create a templated method in the MeshModel and IOMesh to automatically
cast the underlying mesh.

Create an interpolate() method in the VegaVolumetricMesh class to
interpolate surface of the underlying volumetric mesh with an external
surface mesh and clean up the interface.

Create a computeGravity() method in the VegaVolumetricMesh class
to compute gravity forces.

Refactor getVelocity() method in DeformableSceneObject and add a
variable to hold the gravity vector.

Refactor SceneObject. Add SceneModelRenderDelegate render delegate to
take care of renderign operations for this model. Create visual,
physics and collision model variables.

Remove render delegates from the StaticSceneObject, this is now handled
by it base class.

Add storage for the gravity force in VegaFEMDeformableSceneObject and
initialized properly. Compute internal force by multiplying K*positions
instead of calling the vega function.

Refactor DefaultSimulator, VegaFEMModelSimulator. This class will be
marked obsolete soon.