ENH: Refactor VRPN Server/Client API
Allows for the following workflow :
1) Create a `VRPNDeviceClient` by specifying its URL (name@ipadress)
and its type (see `DeviceType` class). The subclass `VRPNForceDevice`
is setting the style to `PHANTOM_OMNI` by default for now.
2) Create a `VRPNDeviceServer`
3) Call `VRPNDeviceServer::addDeviceClient()` to add the client created
on step 1). This will analyse the IP, name, and type of the device
client to instantiate a VRPN device connection. The added devices are
then stored in three `std:map` so the `mainloop()` can be called during
the execution.

Updated their use in `FEMSimulator` and `LaparoscopicCamera` examples.