COMP: Fix for VegaFEM on VS
All projects output directories are now common in
the top level. This commit takes care of the
CMAKE_CONFIGURATION_TYPES for Visual Studio.

With this commit the Find_XXX.cmake files do not need
HINTS anymore since all targets are built in a same
directory known to CMAKE_LIBRARY_PATH, and all headers
directories are referenced in CMAKE_INCLUDE_PATH.