WIP: Implement CameraController
1) Camera does not hold a vtkCamera anymore (less tight integration).
It holds camera parameters.

2) Camera has a pointer to CameraController, that you can set up
based on a deviceclient. This camera controller has a reference
to the camera, and updates its parameters based on the deviceclient
retrieved information.

3) Scene runs the controller as an asynchronous module in its own
thread, it is therefore started in Scene::initModule and ended in
Scene::cleanUpModule.

4) In the timer loop of InteractorStyle, the vtk camera is updated
based on the scene camera.

5) In Renderer, a new vtk Camera needs to be created to match the
scene camera, since it is not created anymore in imstk::Camera.