ENH: Add support for texturing surface mesh
1) In imstkMesh, allow to store point data arrays.

2) In imstkSurfaceMesh, get rid of the API for texture
coordinates and vertice tangents. Instead, we'll store
the name of the data array to use as a default texture
coordinate if none is specified by the user, and offer
the addTexture() API to populate a texture map storing
the filename of the texture and the name of the array
to use as texture coordinates.

3) Update imstkVTKMeshReader to copy all point data,
not just active texture coordinates, and define the
default texture coordinates if active coordinates exist.

4) Update imstkSurfaceMeshRenderDelegate to convert
back point data arrays to vtk data arrays wihtin the
polydata, and use the latest API in VTK to define
multiple texture coordinates and textures.