COMP: Update to specific git_tags for external projects
Using a git branch name for an external project would
update the repository of that external project when
building iMSTK even though it would not be tested.

By hardcoding a certain commit #, we can make sure we
need to specifically update the external project git_tag
to update the external project.

This commit updates to the newer version of VegaFEM which
should resolve some errors:

https://gitlab.kitware.com/iMSTK/VegaFEM-CMake/commit/d1ab0858675d9dbc64345038b3e25cac19b99bbe