ENH: remove interface for transforming geometry in RigidObject
We are still wondering how to store/apply our geometries.
Having yet another interface to only wrap the API to
transform the geometry is not needed: we can directly
call them on the geometry themselves for more transparency
on what geometry is being updated.

This will be investigated later on when we will work with
rigib body dynamics.