ENH: Refactor VirtualCouplingObject in a controller
SceneObjectController: similar to cameraController, it
subclasses trackingController to update the transformation
of an object master geometry, instead of updating itself.

SceneObject: similar to camera, has an interface to easily
setup a controller using a device client.

CollidingObject: stores the force vector that is to be
applied to the device.

SceneManager: make the API calls from controller instead
of object