ENH: Use vtkTriangleMeshPointNormals
The main bottleneck with real-time VTK rendering so far
appeared to be the normals computation. This commit addresses
that issue by making use of the latest normals computation
filter in VTK: vtkTriangleMeshPointNormals.

While this filter is much (5 to 16 times) faster than
vtkPolyDataNormals, it does not check for consistency in the
cell orientations that could cause inverted normals, which
is why the vtkPolyDataNormals is called once in the surfacemesh
renderdelegate to retrieve consistent cells for the input mesh.

See VTK merge request for more information :
https://gitlab.kitware.com/vtk/vtk/merge_requests/2271

PS: That MR requires the latest commits from VTK master, which
does not include work made on texture wrap mode nor on multi
texture attributes yet:
- https://gitlab.kitware.com/iMSTK/vtk/commit/62a7ecd8a5f54e243c26960de22d5d1d23ef932b
- https://gitlab.kitware.com/iMSTK/vtk/commit/ae373026755db42b6fdce5093109ef1a39a76340