ENH: Allow debug rendering while simulation runs
Addresses suggestions made here:
https://gitlab.kitware.com/alexis-girault/iMSTK/commit/a63f98b5de021de37b47521b44b2b9a3df9a2848#note_212643

>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8
It would be great to have the option of using pan-zoom-rotate
when simulation is running for certain applications. I keep
pressing 'S' and 'E' in succession to see if everything is good
for an example i was working on. It need not be a default run
mode option.
>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8>8

* Rename VTKRenderer::setup into VTKRenderer::setMode
* Add VTKRenderer::getMode
* Add VTKViewer::getRenderingMode wrapping the one above
* Consider Rendering Mode in Mouse event callbacks instead of
simulation status
* Allow to switch to Debug rendering and back with the `d` key