COMP: Fix function is not a member of std (Win)
The merge request !117 to add custom event handlers
for vtk interaction introduced a bug on Windows, where
the <functional> header needs to be included to be able
to use `std::function`.

This fixes the 50 errors listed on cdash (due to those
`std::function` members being used frequently):
http://my.cdash.org/viewBuildError.php?buildid=1111691

PS: The experimental build for this MR did not show any
issues on CDash [1], but they appeared once merged in the
master branch (continuous). There seem to be an issue
with how buildbot runs those builds on Windows, since a
similar issue happened on windows 32 bits with Eigen
aligned data structures in stl containers.

[1] http://my.cdash.org/buildSummary.php?buildid=1111615