ENH: default DATA_ROOT_PATH and alert if not set
DATA_ROOT_PATH is used to quickly define the root path
for all your ressources in the Sandbox example. Its value
will therefore vary for every user, which is why it should
not be commited.

This commit adds comments explaining not to commit changes
to this line, as well as how to commit part of a file only,
and displays an error when the path is not set locally.