ENH: Display framerate in window
Press the `P` key (for "Print text", since f should be for
fullscreen) to display the rendering framerate in the left
bottom corner.

Needed to override `SetCurrentRenderer` to add the textActor.

Only update the displayed framerate value every 100ms, which
requires the variables `m_lastFpsUpdate` (last time since the
value was updated), and `m_lastFps` (to compute the framerate
value using a moving average).