ENH: Refactor Tracking Controller
* Rename TrackingController to DeviceTracker
* Have subclasses store a pointer to DeviceTracker instead
of inheriting it. This allows:
1. swapping the tracking controller at runtime
2. having multiple controllers control for the same scene object (eg: laparoscopic tool jaw)
* Add SceneObjectControllerBase to parent SceneObjectController
and other future object controllers
* Update tracking data only once per frame: if the same tracker
is used for multiple objects, it is updated only once per frame.
Each tracker is set to upToDate once the tracking data is updated
and offsets computed once at each frame. At the end of the frame
all the device trackers are set to be outOfDate.