TEST: Add test for imstkPlane
Tests failing:
setgetWidth() with a width <=0
setgetNormal(): the vector returned by getNormal() is slightly different
from the one sent. Indeed, the normal is computed by the difference between
the orientation and the UP_VECTOR. In order to fix the issue, a m_normal
parameter must be added in the imstkPlane class.
Don't forget to update the rendering part based on this new m_normal vector