ENH: Allow custom library definition
Expand the imstk_add_library cmake function by exposing the variables
H_FILES (headers) CPP_FILES (source) and SUBDIR_LIST (build interface
directories - the current source dir is always included).

If those values are not manually set, the previous behavior applies,
where all the files and directories in the root directory are used for
that target.

Used that new mechanism for the Rendering target to prepare for Vulkan
integration.