ENH: Adds capsule-mesh collision detection
Adds capsule-point cloud collision detection.
First the each point in mesh is checked against
the bounding sphere of the capsule. If inside,
the closest point is found on the line joining
the two apexes of the capsule. If the point
is inside the bounding sphere then the closest
point should be inside the capsule. If the distance
between the closest point and the mesh point is less
than radius of the hemispheres of the capsule, the
collision is registered.