BUG: Fix #122 black analytic objects
Fix regression introduced in !121 with the use of the fast normals filter
`vtkTriangleMeshPointNormals`. That filter was also applied for surface
meshes with non-triangular cells: cells, sphere, capsule, cube (quads and
not triangles).

Since the issue of non-triangular meshes comes from analytic mesh sources,
the output surface meshes will always be rigid, so their normals will stay
constant and won't need to be recomputed. We can therefore use the filter
`vtkPolyDataNormals`, which is slower but will only need to be called once
of no deformation occurs.