ENH: Use UPS through lambdas
The current integration of UPS within the module class did not allow
to display the UPS value in any easy and standardized way.

This commits simplifies the Module class by removing its dependency to
the UPSCounter class, and shows how to easily query the UPS using the
callback mechanism on any module (here: sceneManager for pbd physics and
fem physics examples).