ENH: Collision handling for Edge-Edge CCD
Add collision handling for Edge-Edge transient collision.
Add step to cache resolved collision geometries to be used as "previous
time-step" for next iteration.

Fix indicesPtr bug in MeshSide. Previously nullptr was passed to it by
default in all constructions. Add getCellIndices() function to PointSet
to populate the pointer for MeshSide construction if available.