ENH: Replace all references to CollidingObject with Collider equivalent
Remove CollidingObject class.
Replace all mentions of CollidingObject class with either Entity
or SceneObject as deemed appropriate.

Replace CollidingObject::getCollidingGeometry() calls with
Entity::getComponent<Collider>().getGeometry(). There is a temporary
built-in assumption that there is a single collider associated with
an Entity. Will need to revisit this at some point for multi-part
/ multi-collider entities.