PERF: Pre-fetch required Components at initialization
Pre-fetch and place required components such as colliders
into member variables during initialization,
rather than fetching them in update or inside the simulation
loop. Entity::getComponent<T> iterates over a vector of
Components and checks every item by doing a dynamic_cast.
This is an expensive operation and should be avoided within
the simulation loop.

NOTE: This is a partial fix, since there are many other instances that
fetch components from entities that are called within the simulation
loop such as update() and handle() functions of interaction and
collision handling classes. This should be addressed in a separate
more holistic merge request.